+++ lib/vagrant-lxc/driver/cli.rb
@@ -10,6 +10,7 @@ module Vagrant
         attr_accessor :name
 
         class TransitionBlockNotProvided < RuntimeError; end
+        class ShutdownNotSupported < RuntimeError; end
         class TargetStateNotReached < RuntimeError
           def initialize(target_state, state)
             msg = "Target state '#{target_state}' not reached, currently on '#{state}'"
@@ -72,7 +73,12 @@ module Vagrant
         end
 
         def shutdown
-          run :shutdown, '--name', @name
+          if system('which lxc-shutdown > /dev/null')
+            run :shutdown, '--name', @name
+          else
+            # REFACTOR: Do not use exception to control the flow
+            raise ShutdownNotSupported
+          end
         end
 
         def attach(*cmd)