+++ spec/unit/driver/cli_spec.rb
@@ -103,12 +103,18 @@ describe Vagrant::LXC::Driver::CLI do
     subject    { described_class.new(sudo_wrapper, name) }
 
     before do
+      subject.stub(system: true)
       subject.stub(:run)
-      subject.shutdown
     end
 
     it 'issues a lxc-shutdown with provided container name' do
+      subject.shutdown
       subject.should have_received(:run).with(:shutdown, '--name', name)
+    end
+
+    it 'raises a ShutdownNotSupported in case it is not supported' do
+      subject.stub(:system).with('which lxc-shutdown > /dev/null').and_return(false)
+      expect { subject.shutdown }.to raise_error(described_class::ShutdownNotSupported)
     end
   end