+++ spec/unit/driver_spec.rb
@@ -118,6 +118,13 @@ describe Vagrant::LXC::Driver do
       cli.should_receive(:stop)
       subject.forced_halt
     end
+
+    it 'attempts to force the container to stop in case lxc-shutdown is not supported' do
+      cli.stub(:shutdown).and_raise(Vagrant::LXC::Driver::CLI::ShutdownNotSupported)
+      cli.should_receive(:transition_to).with(:stopped).twice
+      cli.should_receive(:stop)
+      subject.forced_halt
+    end
   end
 
   describe 'state' do